Adding database config into its own file
