+++ controllers/authController.js
@@ -1,7 +1,6 @@
 const bcrypt = require("bcrypt");
 const jwt = require("jsonwebtoken");
-const sequelize = require("../config/config");
-const db = require("../config/config");
+const sequelize = require("../config/database");
 const validateEmail = require("../utils/emailValidator");
 const User = require("../models/User");
 const Driver = require("../models/Driver");
