+++ models/Driver.js
@@ -1,5 +1,5 @@
 const { DataTypes } = require("sequelize");
-const sequelize = require("../config/config");
+const sequelize = require("../config/database");
 const User = require("./User");
 
 const Driver = sequelize.define(
