+++ models/Job.js
@@ -1,5 +1,5 @@
 const { DataTypes } = require("sequelize");
-const sequelize = require("../config/config");
+const sequelize = require("../config/database");
 const Driver = require("./Driver");
 const Shipper = require("./Shipper");
