+++ models/User.js
@@ -1,4 +1,4 @@
-const sequelize = require("../config/config");
+const sequelize = require("../config/database");
 const { DataTypes } = require("sequelize");
 
 const User = sequelize.define(
