+++ models/Vehicle.js
@@ -1,5 +1,5 @@
 const { DataTypes } = require("sequelize");
-const sequelize = require("../config/config");
+const sequelize = require("../config/database");
 const Driver = require("./Driver");
 
 const Vehicle = sequelize.define(
